feat: add migration and reshare flows with new components for secure vault
