+++ task.md
@@ -1,3 +1,7 @@
 # Notes for the task related to the current pull request.
 
 The content of this file should be cleared on Pull Request submit.
+
+https://github.com/vultisig/vultisig-windows/issues/1281
+
+- [ ] Share fast reshare and migrate flows
